[backend] reorder - add db/ to all db controller - add modules/user.[service|module|controller].ts   - this add /user/me route

+++ src/nestjs/src/modules/database/gameInfo/service.ts
@@ -1,4 +1,4 @@
-import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
+import { Injectable, ForbiddenException } from "@nestjs/common";
 import { Repository } from "typeorm";
 import { InjectRepository } from "@nestjs/typeorm";
 
@@ -9,54 +9,44 @@ import { UserEntity } from "../user/entity";
 
 @Injectable()
 export class GameInfoService {
-        constructor(
-                @InjectRepository(GameInfoEntity)
-                private readonly gameInfoRepo: Repository<GameInfoEntity>,
-                @InjectRepository(UserEntity)
-                private readonly userRepo: Repository<UserEntity>,
-        ) {}
+	constructor(
+		@InjectRepository(GameInfoEntity)
+		private readonly gameInfoRepo: Repository<GameInfoEntity>,
+		@InjectRepository(UserEntity)
+		private readonly userRepo: Repository<UserEntity>,
+	) {}
 
-        async create(post: GameInfoPost, userA: number, userB: number) {
+	async create(post: GameInfoPost, userA: number, userB: number) {
+		const gameInfo = new GameInfoEntity();
+		const user1 = await this.userRepo.findOneBy({ id: userA });
+		const user2 = await this.userRepo.findOneBy({ id: userB });
+		if (user1 && user2) {
+			gameInfo.userA = userA;
+			gameInfo.userB = userB;
+			gameInfo.type = post.type;
+			return await this.gameInfoRepo.save(gameInfo);
+		} else throw new ForbiddenException("User not found");
+	}
 
-                const gameInfo = new GameInfoEntity();
-                const user1 = await this.userRepo.findOneBy({id: userA}); 
-                const user2 = await this.userRepo.findOneBy({id: userB}); 
-                if (user1 && user2)
-                {
-                        gameInfo.userA = userA;
-                        gameInfo.userB = userB;
-                        gameInfo.type = post.type;
-                        return await this.gameInfoRepo.save(gameInfo);
-                }
-                else
-                        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
-          }
+	async returnAll() {
+		return await this.gameInfoRepo.find();
+	}
 
-        async returnAll() {
-                return await this.gameInfoRepo.find();
-        }
+	async returnOne(gameId: number) {
+		const tmp = await this.gameInfoRepo.findOneBy({ id: gameId });
+		if (tmp) return await this.gameInfoRepo.findOneBy({ id: gameId });
+		else throw new ForbiddenException("GameInfo not found");
+	}
 
-        async returnOne(gameId: number) {
-                const tmp = await this.gameInfoRepo.findOneBy({id: gameId});
-                if (tmp)
-                        return await this.gameInfoRepo.findOneBy({ id: gameId });
-                else
-                        throw new HttpException('GameInfo not found', HttpStatus.NOT_FOUND);
-        }
+	async update(gameId: number, post: GameInfoPost) {
+		const tmp = await this.gameInfoRepo.findOneBy({ id: gameId });
+		if (tmp) return await this.gameInfoRepo.update(gameId, post);
+		else throw new ForbiddenException("GameInfo not found");
+	}
 
-        async update(gameId: number, post: GameInfoPost) {
-                const tmp = await this.gameInfoRepo.findOneBy({id: gameId});
-                if (tmp)
-                        return await this.gameInfoRepo.update(gameId, post);
-                else
-                        throw new HttpException('GameInfo not found', HttpStatus.NOT_FOUND);
-        }
-
-        async delete(gameId: number) {
-                const tmp = await this.gameInfoRepo.findOneBy({ id: gameId});
-                if (tmp)
-                        return await this.gameInfoRepo.delete(tmp);
-                else
-                        throw new HttpException('GameInfo not found', HttpStatus.NOT_FOUND);
-        }
+	async delete(gameId: number) {
+		const tmp = await this.gameInfoRepo.findOneBy({ id: gameId });
+		if (tmp) return await this.gameInfoRepo.delete(tmp);
+		else throw new ForbiddenException("GameInfo not found");
+	}
 }
